move SignatureAlgorithm to `types.ts`
